feat(Multiple)Feign Client Implementation

+++ customer_service/src/main/java/com/onlineshop/customer_service/service/CustomerService.java
@@ -2,6 +2,7 @@ package com.onlineshop.customer_service.service;
 
 import java.util.List;
 
+import com.onlineshop.customer_service.client.OrderServiceClient;
 import com.onlineshop.customer_service.entity.Customer;
 import com.onlineshop.customer_service.exception.CustomerIdMismatchException;
 import com.onlineshop.customer_service.exception.CustomerNotFoundException;
@@ -14,10 +15,13 @@ import org.springframework.stereotype.Service;
 public class CustomerService {
 
     private final CustomerRepository customerRepository;
+    private final OrderServiceClient orderClient;
 
     @Autowired
-    public CustomerService(CustomerRepository customerRepository) {
+    public CustomerService(CustomerRepository customerRepository,
+                           OrderServiceClient orderClient) {
         this.customerRepository = customerRepository;
+        this.orderClient = orderClient;
     }
 
     // fetch all customers
@@ -40,9 +44,13 @@ public class CustomerService {
     // delete customers
     // TODO: ADD EXCEPTION IF ORDERS EXIST
     public void delete(long id){
-        customerRepository.findById(id)
+        if (!checkOrders(id)){
+            customerRepository.findById(id)
                 .orElseThrow(CustomerNotFoundException::new);
-        customerRepository.deleteById(id);
+            customerRepository.deleteById(id);
+        } else {
+            System.out.println("There are still open orders on this customer!");
+        }
     }
 
     // update customers
@@ -53,5 +61,10 @@ public class CustomerService {
         customerRepository.findById(id)
                 .orElseThrow(CustomerNotFoundException::new);
         return customerRepository.save(customer);
+    }
+
+    // check if there are open orders
+    public boolean checkOrders(Long customerId){
+        return this.orderClient.getCustomerOrders(customerId);
     }
 }
